+++ mayan/apps/file_metadata/tests/mixins/document_file_mixins.py
@@ -1,31 +1,10 @@
-from django.utils.module_loading import import_string
-
-from mayan.apps.documents.tests.mixins.document_mixins import DocumentTestMixin
-
-from ..classes import FileMetadataDriver
-
-from .literals import (
-    TEST_DRIVER_CLASS_PATH, TEST_FILE_METADATA_KEY, TEST_FILE_METADATA_VALUE
+from mayan.apps.documents.tests.mixins.document_mixins import (
+    DocumentTestMixin
 )
 
+from ..literals import TEST_FILE_METADATA_KEY, TEST_FILE_METADATA_VALUE
 
-class FileMetadataTestMixin:
-    _test_document_file_metadata_driver_path = TEST_DRIVER_CLASS_PATH
-
-    def setUp(self):
-        super().setUp()
-        FileMetadataDriver.load_modules()
-
-        FileMetadataDriver.collection.do_driver_disable_all()
-
-        if self._test_document_file_metadata_driver_path:
-            self._test_document_file_metadata_driver = import_string(
-                dotted_path=self._test_document_file_metadata_driver_path
-            )
-            self._test_document_file_metadata_driver.do_model_instance_populate()
-            FileMetadataDriver.collection.do_driver_enable(
-                driver=self._test_document_file_metadata_driver
-            )
+from .file_metadata_mixins import FileMetadataTestMixin
 
 
 class DocumentFileMetadataTestMixin(
@@ -83,23 +62,3 @@ class DocumentFileMetadataViewTestMixin(DocumentFileMetadataTestMixin):
                 'id_list': self._test_document_file.pk
             }
         )
-
-
-class DocumentTypeViewTestMixin(FileMetadataTestMixin):
-    def _request_document_type_file_metadata_settings_view(self):
-        return self.get(
-            viewname='file_metadata:document_type_file_metadata_settings',
-            kwargs={'document_type_id': self._test_document.document_type.pk}
-        )
-
-    def _request_document_type_file_metadata_submit_view(self):
-        return self.post(
-            viewname='file_metadata:document_type_file_metadata_submit', data={
-                'document_type': self._test_document_type.pk,
-            }
-        )
-
-
-class FileMetadataDriverTestMixin(FileMetadataTestMixin):
-    def _request_file_metadata_driver_list_view(self):
-        return self.get(viewname='file_metadata:file_metadata_driver_list')