+++ mayan/apps/file_metadata_clamav/tests/test_drivers.py
@@ -1,5 +1,7 @@
 from mayan.apps.documents.tests.base import GenericDocumentTestCase
-from mayan.apps.file_metadata.tests.mixins import DocumentFileMetadataTestMixin
+from mayan.apps.file_metadata.tests.mixins.document_file_mixins import (
+    DocumentFileMetadataTestMixin
+)
 
 from ..drivers import ClamScanDriver